requirements.txt
asgiref==3.7.2
Django==4.2.5
django-cleanup==8.0.0
django-cors-headers==4.2.0
django-filter==23.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
et-xmlfile==1.1.0
numpy==1.26.0
openpyxl==3.1.2
pandas==2.1.0
Pillow==10.0.1
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
whitenoise==6.5.0
